requirements.txt
matplotlib==3.5.1
scikit-learn==1.0.2
tensorflow==2.8.0
mlflow==1.26.0
librosa==0.9.1
protobuf==3.20.0
pyyaml==6.0
dvc==2.8.1
psycopg2==2.9.3

pytest==7.1.2